+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.7.0-beta.5
 ## mm/dd/2019
 
+1. [](#new)
+    * Added a new `bin/grav server` CLI command to easily run Symfony or PHP built-in webservers
 1. [](#improved)
     * Better support for Symfony local server `symfony server:start`
     * Make `Route` objects immutable